Fix Ticket 6605

+++ Controllers/TransitoTransporteController.cs
@@ -12,6 +12,7 @@ using System.Net;
 using System.IO;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
+using Microsoft.IdentityModel.Tokens;
 
 namespace GuanajuatoAdminUsuarios.Controllers
 {
@@ -139,6 +140,12 @@ namespace GuanajuatoAdminUsuarios.Controllers
         [HttpPost]
         public ActionResult ajax_DeleteTransito(string ids)
         {
+            if (ids.IsNullOrEmpty())
+            {
+                Response.StatusCode = (int)HttpStatusCode.BadRequest;
+                return null;
+            }                
+
             string[] idsList = ids.Split(",");
             var result = _transitoTransporteService.DeleteTransitoTransporte(Convert.ToInt32(idsList[0]), Convert.ToInt32(idsList[1]));
             if (result > 0)
